feat(memory): Add ZFS ARC (cache) calculation to memory module (#21)
* Added zfs arc calculation

Redesigned the way 32-memory gets memory stats to make it more precise when subtracting the size of the zfs arc. If ZFS is detected on the running system, it will subtract the size of its cache to provide a more accurate output.

* Updated memory unit formatting

The memory units were being converted to mb, subtracted than converted to gb. On low-memory machines (under 1G) it would only display .3G instead of the more readable 300M. By using numfmt, this behavior is restored :)

* Updated comments